@@ -3,86 +3,37 @@
          xmlns="http://maven.apache.org/POM/4.0.0"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-
-    <groupId>org.shiminfxcvii</groupId>
-    <artifactId>employee-management</artifactId>
-    <version>1.0-SNAPSHOT</version>
-
-    <!-- 18 似乎指的是 JDK 版本 -->
-    <!-- 如果改成 19 ====> java: 错误: 不支持发行版本 19 -->
-    <properties>
-        <maven.compiler.source>19</maven.compiler.source>
-        <maven.compiler.target>19</maven.compiler.target>
-
-        <java.version>19</java.version>
-
-        <spring-boot.version>3.0.0-SNAPSHOT</spring-boot.version>
-
-        <spring-framework.version>6.0.0-SNAPSHOT</spring-framework.version>
-
-        <lombok.version>1.18.24</lombok.version>
-
-        <mysql.version>8.0.30</mysql.version>
-
-        <thymeleaf.extras.version>3.1.0.M1</thymeleaf.extras.version>
-
-        <jetbrains.annotations.version>23.0.0</jetbrains.annotations.version>
-
-        <fasterxml.jackson.core.version>2.13.4</fasterxml.jackson.core.version>
-
-        <springfox.version>3.0.0</springfox.version>
-
-        <!-- spring-boot 官方项目的某个配置 -->
-        <!-- <thymeleaf.version>3.0.2.RELEASE</thymeleaf.version>
-		<thymeleaf-layout-dialect.version>2.1.1</thymeleaf-layout-dialect.version> -->
-
-        <!-- 升级到 3.0.15 -->
-        <!--        <thymeleaf.version>3.0.15.RELEASE</thymeleaf.version>-->
-        <!-- 作用：向主页面引入子页面 -->
-        <!--        <thymeleaf-layout-dialect.version>3.0.15</thymeleaf-layout-dialect.version>-->
-    </properties>
-
-    <!-- 父引用 -->
-    <!-- 如果是 spring-boot-parent 是否会报错？ -->
-    <!-- 不会 -->
-    <!-- 2022/4/7 今天从 2.6.4 升级到 2.6.5 -->
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.0.0-SNAPSHOT</version>
-        <relativePath/>
+        <version>3.0.5</version>
+        <relativePath/> <!-- lookup parent from repository -->
     </parent>
 
+    <groupId>com.shiminfxcvii</groupId>
+    <artifactId>employee-management</artifactId>
+    <version>1.0-SNAPSHOT</version>
+
+    <properties>
+        <java.version>20</java.version>
+        <maven.compiler.source>${java.version}</maven.compiler.source>
+        <maven.compiler.target>${java.version}</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <project.reporting.outputEncoding>${project.build.sourceEncoding}</project.reporting.outputEncoding>
+
+        <springdoc.version>2.1.0</springdoc.version>
+        <jetbrains.annotations.version>24.0.1</jetbrains.annotations.version>
+    </properties>
+
     <dependencies>
-
-        <!-- spring-beans & jackson-databind dependencies 是因为报告以下信息而添加的 -->
-        <!-- Provides transitive vulnerable dependency
-         'com.fasterxml.jackson.core':jackson-databind:2.13.2 Out-of-bounds Write vulnerability pending CVSS allocation
-         Provides transitive vulnerable dependency 'org.springframework':spring-beans:5.3.17 Improper Control of
-         Generation of Code ('Code Injection') vulnerability pending CVSS allocation -->
-
-        <!-- Provides transitive vulnerable dependency -->
-        <!-- 2.13.2 Out-of-bounds Write vulnerability pending CVSS allocation -->
-        <!-- Out-Of-Bounds Write -->
-        <!-- Summary
-         jackson-databind before 2.12.6.1 and 2.13.x before 2.13.2.1 allows a Java StackOverflow exception and denial
-         of service via a large depth of nested objects. -->
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <version>${fasterxml.jackson.core.version}</version>
-        </dependency>
-
         <!-- MySQL 驱动 -->
+        <!-- https://mvnrepository.com/artifact/com.mysql/mysql-connector-j -->
         <dependency>
-            <groupId>mysql</groupId>
-            <artifactId>mysql-connector-java</artifactId>
-            <version>${mysql.version}</version>
+            <groupId>com.mysql</groupId>
+            <artifactId>mysql-connector-j</artifactId>
         </dependency>
 
-        <!-- 不知道有什么作用，是有一次编写代码使提示添加的 -->
-        <!-- 知道了，是为了导入下面的包 -->
-        <!-- import org.jetbrains.annotations.NotNull; -->
+        <!-- jetbrains -->
         <dependency>
             <groupId>org.jetbrains</groupId>
             <artifactId>annotations</artifactId>
