@@ -2,33 +2,23 @@ package com.shiminfxcvii.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 import lombok.Setter;
-import lombok.ToString;
-
-import java.util.Objects;
+import org.springframework.data.jpa.domain.AbstractAuditable;
 
 /**
  * 员工信息实体类
  *
- * @author shiminfxcvii
+ * @author ShiminFXCVII
  * @since 6/2/2022 5:16 PM
  */
 @Entity
 @Getter
 @Setter
-@ToString
-@RequiredArgsConstructor
 @Table(name = "employee", schema = "employee_management")
-public class Employee {
+public class Employee extends AbstractAuditable<Long, Long> {
 
-    @Id
-    @Column(name = "employee_id", unique = true, nullable = false, updatable = false, columnDefinition = "varchar",
-            length = 36)
-    private String employeeId;
     @Column(name = "employee_name", nullable = false, columnDefinition = "varchar", length = 45)
     private String employeeName;
     @Column(name = "employee_sex", nullable = false, columnDefinition = "varchar", length = 1)
@@ -41,24 +31,5 @@ public class Employee {
     private String employeeAddress;
     @Column(name = "employee_phone_number", nullable = false, columnDefinition = "varchar", length = 11)
     private String employeePhoneNumber;
-    @Column(name = "created_by", nullable = false, updatable = false, columnDefinition = "varchar", length = 45)
-    private String createdBy;
-    @Column(name = "created_date", nullable = false, updatable = false, columnDefinition = "varchar", length = 19)
-    private String createdDate;
-    @Column(name = "last_modified_date", nullable = false, columnDefinition = "varchar", length = 19)
-    private String lastModifiedDate;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (null == o || getClass() != o.getClass()) return false;
-        Employee employee = (Employee) o;
-        return null != employeeId && Objects.equals(employeeId, employee.employeeId);
-    }
-
-    @Override
-    public int hashCode() {
-        return getClass().hashCode();
-    }
 
 }
